@@ -5,27 +5,21 @@ from androguard import __version__
 from setuptools import setup, find_packages
 
 
-# We do not support Python <3.4
-if sys.version_info < (3, 4):
+# We do not support Python <3.5 (lxml and pyqt5 are not supported as well)
+if sys.version_info < (3, 5):
     print("Unfortunately, your python version is not supported!\n"
-          "Please upgrade at least to Python 3.4!", file=sys.stderr)
+          "Please upgrade at least to Python 3.5!", file=sys.stderr)
     sys.exit(1)
 
-# PyQT5 is only available for python >=3.5
-if sys.version_info <= (3, 4):
-    print("PyQT5 is probably not available for your system, the GUI might not work!", file=sys.stderr)
+with open('requirements.txt', 'r') as fp:
+    install_requires = fp.read().splitlines()
 
-install_requires = [
-                    'networkx>=1.11',
-                    'pygments',
-                    'lxml',
-                    'colorama',
-                    'matplotlib',
-                    'asn1crypto>=0.24.0',
-                    'click',
-                    'pydot>=1.4.1',
-                    'ipython>=5.0.0',
-                    ]
+# Find the right version for the magic package
+if sys.platform in ('darwin', 'win32'):
+    magic_package = 'python-magic-bin>=0.4.14'
+else:
+    magic_package = 'python-magic>=0.4.15'
+
 
 # TODO add the permission mapping generation at a better place!
 # from axplorer_to_androguard import generate_mappings
@@ -78,17 +72,16 @@ setup(
     install_requires=install_requires,
     extras_require={
         'GUI': ["pyperclip", "PyQt5"],
-        'magic': ['python-magic>=0.4.15'],
+        'magic': [magic_package],
         'docs': ['sphinx', "sphinxcontrib-programoutput>0.8", 'sphinx_rtd_theme'],
         'tests': ['mock>=2.0', 'nose', 'codecov', 'coverage', 'nose-timer'],
     },
     setup_requires=['setuptools'],
-    python_requires='>=3.4',
+    python_requires='>=3.5',
     classifiers=[
                  'License :: OSI Approved :: Apache Software License',
                  'Programming Language :: Python',
                  'Programming Language :: Python :: 3',
-                 'Programming Language :: Python :: 3.4',
                  'Programming Language :: Python :: 3.5',
                  'Programming Language :: Python :: 3.6',
                  'Programming Language :: Python :: 3.7',
