@@ -48,8 +48,7 @@ logger = logging.getLogger('dad')
 
 # No seperate DvField class currently
 def get_field_ast(field):
-    triple = field.get_class_name()[1:-1], field.get_name(
-    ), field.get_descriptor()
+    triple = field.get_class_name()[1:-1], field.get_name(), field.get_descriptor()
 
     expr = None
     if field.init_value:
@@ -74,11 +73,10 @@ class DvMethod:
     """
     This is a wrapper around :class:`~androguard.core.analysis.analysis.MethodAnalysis` and
     :class:`~androguard.core.bytecodes.dvm.EncodedMethod` inside the decompiler.
+
+    :param androguard.core.analysis.analysis.MethodAnalysis methanalysis:
     """
     def __init__(self, methanalysis):
-        """
-        :param androguard.core.analysis.analysis.MethodAnalysis methanalysis:
-        """
         method = methanalysis.get_method()
         self.method = method
         self.start_block = next(methanalysis.get_basic_blocks().get(), None)
@@ -121,6 +119,19 @@ class DvMethod:
             bytecode.method2png('/tmp/dad/graphs/{}#{}.png'.format(self.cls_name.split('/')[-1][:-1], self.name), methanalysis)
 
     def process(self, doAST=False):
+        """
+        Processes the method and decompile the code.
+
+        There are two modes of operation:
+
+        1) Normal Decompilation to Java Code
+        2) Decompilation into an abstract syntax tree (AST)
+
+        The Decompilation is done twice. First, a rough decompilation is created,
+        which is then optimized. Second, the optimized version is used to create the final version.
+
+        :param doAST: generate AST instead of Java Code
+        """
         logger.debug('METHOD : %s', self.name)
 
         # Native methods... no blocks.
@@ -182,6 +193,22 @@ class DvMethod:
             self.writer.write_method()
 
     def get_ast(self):
+        """
+        Returns the AST, if previously was generated by calling :meth:`process` with argument :code:`doAST=True`.
+
+        The AST is a :class:`dict` with the following keys:
+
+        * triple
+        * flags
+        * ret
+        * params
+        * comments
+        * body
+
+        The actual AST for the method is in the :code:`body`.
+
+        :return: dict
+        """
         return self.ast
 
     def show_source(self):
@@ -199,7 +226,7 @@ class DvMethod:
 
     def __repr__(self):
         # return 'Method %s' % self.name
-        return 'class DvMethod(object): %s' % self.name
+        return '<class DvMethod(object): %s>' % self.name
 
 
 class DvClass:
@@ -208,12 +235,11 @@ class DvClass:
 
     At first, :py:attr:`methods` contains a list of :class:`~androguard.core.bytecodes.dvm.EncodedMethods`,
     which are successively replaced by :class:`DvMethod` in the process of decompilation.
+
+    :param androguard.core.bytecodes.dvm.ClassDefItem dvclass: the class item
+    :param androguard.core.analysis.analysis.Analysis vma: an Analysis object
     """
     def __init__(self, dvclass, vma):
-        """
-        :param androguard.core.bytecodes.dvm.ClassDefItem dvclass: the class item
-        :param androguard.core.analysis.analysis.Analysis vma: an Analysis object
-        """
         name = dvclass.get_name()
         if name.find('/') > 0:
             pckg, name = name.rsplit('/', 1)
@@ -299,7 +325,7 @@ class DvClass:
 
         if len(self.interfaces) > 0:
             prototype += ' implements %s' % ', '.join(
-                [n[1:-1].replace('/', '.') for n in self.interfaces])
+                [str(n[1:-1].replace('/', '.')) for n in self.interfaces])
 
         source.append('%s {\n' % prototype)
         for field in self.fields:
@@ -315,7 +341,7 @@ class DvClass:
                 value = init_value.value
                 if f_type == 'String':
                     if value:
-                        value = '"%s"' % value.encode("unicode-escape").decode("ascii")
+                        value = '"%s"' % str(value).encode("unicode-escape").decode("ascii")
                     else:
                         # FIXME we can not check if this value here is null or ""
                         # In both cases we end up here...
@@ -411,7 +437,7 @@ class DvClass:
         print(self.get_source())
 
     def __repr__(self):
-        return 'Class(%s)' % self.name
+        return '<Class(%s)>' % self.name
 
 
 class DvMachine:
