@@ -1,4 +1,4 @@
 # The current version of Androguard
 # Please use only this variable in any scripts,
 # to keep the version number the same everywhere.
-__version__ = "3.4.0"
+__version__ = "3.4.0a1"
