@@ -45,7 +45,7 @@ def entry_point(verbosity):
 @entry_point.command()
 @click.option(
     '--input', '-i', 'input_',
-    type=click.Path(exists=True),
+    type=click.Path(exists=True, file_okay=True, dir_okay=False),
     help='AndroidManifest.xml or APK to parse (legacy option)',
 )
 @click.option(
@@ -53,13 +53,12 @@ def entry_point(verbosity):
     help='filename to save the decoded AndroidManifest.xml to, default stdout',
 )
 @click.option("--resource", "-r",
-        help="Resource inside the APK to parse instead of AndroidManifest.xml"
+        help="Resource (any binary XML file) inside the APK to parse instead of AndroidManifest.xml"
 )
 @click.argument(
     'file_',
-    type=click.Path(exists=True),
-    # help='AndroidManifest.xml or APK to parse',
     required=False,
+    type=click.Path(exists=True, file_okay=True, dir_okay=False),
 )
 def axml(input_, output, file_, resource):
     """
@@ -99,7 +98,6 @@ def axml(input_, output, file_, resource):
 )
 @click.argument(
     'file_',
-    # help='resources.arsc or APK to parse',
     required=False,
 )
 @click.option(
@@ -125,17 +123,17 @@ def axml(input_, output, file_, resource):
     help="Resolve the given ID for the given locale and package. Provide the hex ID!"
 )
 @click.option(
-    '--list-packages', '-t', is_flag=True,
+    '--list-packages', is_flag=True,
     default=False,
     help='List all package names and exit',
 )
 @click.option(
-    '--list-locales', '-t', is_flag=True,
+    '--list-locales', is_flag=True,
     default=False,
     help='List all package names and exit',
 )
 @click.option(
-    '--list-types', '-t', is_flag=True,
+    '--list-types', is_flag=True,
     default=False,
     help='List all types and exit',
 )
@@ -252,11 +250,11 @@ def arsc(input_,
     '--output', '-o',
     default="callgraph.gml", show_default=True,
     help='Filename of the output file, the extension is used to decide which '
-         'format to use (default callgraph.gml)',
+         'format to use',
 )
 @click.option(
     '--show', '-s',
-    default=False,
+    is_flag=True,
     help='instead of saving the graph, print it with mathplotlib '
          '(you might not see anything!)',
 )
@@ -292,10 +290,12 @@ def arsc(input_,
 )
 @click.argument(
     'APK',
-    # help='The APK to analyze',
     nargs=1,
-    required=False,
-    type=click.Path(exists=True),
+    type=click.Path(exists=True,
+                    file_okay=True,
+                    dir_okay=False,
+                    readable=True,
+                    allow_dash=False),
 )
 def cg(output,
        show,
@@ -309,12 +309,15 @@ def cg(output,
     """
     Create a call graph and export it into a graph format.
 
+    The default is to create a file called callgraph.gml in the current
+    directory!
+
     classnames are found in the type "Lfoo/bar/bla;".
 
     Example:
 
     \b
-        $ androguard cg APK
+        $ androguard cg examples/tests/hello-world.apk
     """
     androcg_main(verbose=verbose,
                  APK=apk,
@@ -330,13 +333,12 @@ def cg(output,
 @entry_point.command()
 @click.option(
     '--input', '-i', 'input_',
-    type=click.Path(exists=True),
+    type=click.Path(exists=True, dir_okay=False, file_okay=True),
     help='APK to parse (legacy option)',
 )
 @click.argument(
     'file_',
-    type=click.Path(exists=True),
-    # help='APK to parse',
+    type=click.Path(exists=True, dir_okay=False, file_okay=True),
     required=False,
 )
 @click.option(
@@ -417,10 +419,8 @@ def decompile(input_, file_, output, format_, jar, limit, decompiler):
 )
 @click.argument(
     'apk',
-    # help='APK(s) to extract the Fingerprint of Certificates from',
     nargs=-1,
-    required=False,
-    type=click.Path(exists=True),
+    type=click.Path(exists=True, dir_okay=False, file_okay=True),
 )
 def sign(hash_, print_all_hashes, show, apk):
     """Return the fingerprint(s) of all certificates inside an APK."""
@@ -431,8 +431,7 @@ def sign(hash_, print_all_hashes, show, apk):
 @click.argument(
     'apks',
     nargs=-1,
-    required=False,
-    type=click.Path(exists=True),
+    type=click.Path(exists=True, file_okay=True, dir_okay=False),
 )
 def apkid(apks):
     """Return the packageName/versionCode/versionName per APK as JSON."""
@@ -448,10 +447,12 @@ def apkid(apks):
 @entry_point.command()
 @click.option(
     '--input_file', '-i',
-    type=click.Path(exists=True),
+    help="Specify the inital file to load in the GUI",
+    type=click.Path(exists=True, dir_okay=False, file_okay=True),
 )
 @click.option(
     '--input_plugin', '-p',
+    help="Additional Plugin (currently unused)",
     type=click.Path(exists=True),
 )
 def gui(input_file, input_plugin):
@@ -467,10 +468,9 @@ def gui(input_file, input_plugin):
 )
 @click.argument(
     'apk',
-    # help='Start the shell with the given APK. a, d, dx are available then. Loading might be slower in this case!',
     default=None,
     required=False,
-    type=click.Path(exists=True),
+    type=click.Path(exists=True, dir_okay=False, file_okay=True),
 )
 def analyze(session, apk):
     """Open a IPython Shell and start reverse engineering."""
@@ -486,7 +486,10 @@ def analyze(session, apk):
         default=0,
         type=int,
         help="Number of bytes from offset to disassemble, 0 for whole file")
-@click.argument("DEX")
+@click.argument(
+        "DEX",
+        type=click.Path(exists=True, dir_okay=False, file_okay=True),
+        )
 def disassemble(offset, size, dex):
     """
     Disassemble Dalvik Code with size SIZE starting from an offset
@@ -496,3 +499,4 @@ def disassemble(offset, size, dex):
 
 if __name__ == '__main__':
     entry_point()
+
