@@ -1,19 +1,22 @@
 import re
 import collections
+from operator import itemgetter
 import time
-import warnings
 from androguard.core.androconf import is_ascii_problem, load_api_specific_resource_module
 from androguard.core.bytecodes import dvm
 import logging
-from androguard.core import bytecode
+from androguard.core import bytecode, mutf8
 import networkx as nx
 from enum import IntEnum
 
 log = logging.getLogger("androguard.analysis")
 
-BasicOPCODES = []
+BasicOPCODES = set()
 for i in dvm.BRANCH_DVM_OPCODES:
-    BasicOPCODES.append(re.compile(i))
+    p = re.compile(i)
+    for op, items in dvm.DALVIK_OPCODES_FORMAT.items():
+        if p.match(items[1][0]):
+            BasicOPCODES.add(op)
 
 
 class REF_TYPE(IntEnum):
@@ -38,9 +41,11 @@ class REF_TYPE(IntEnum):
 
 class DVMBasicBlock:
     """
-        A simple basic block of a dalvik method
-    """
+    A simple basic block of a dalvik method.
 
+    A basic block consists of a series of :class:`~androguard.core.bytecodes.dvm.Instruction`
+    which are not interrupted by branch or jump instructions such as `goto`, `if`, `throw`, `return`, `switch` etc.
+    """
     def __init__(self, start, vm, method, context):
         self.__vm = vm
         self.method = method
@@ -57,7 +62,7 @@ class DVMBasicBlock:
 
         self.special_ins = {}
 
-        self.name = "{}-BB@0x{:x}".format(self.method.get_name(), self.start)
+        self.name = mutf8.MUTF8String.join([self.method.get_name(), b'-BB@', hex(self.start).encode()])
         self.exception_analysis = None
 
         self.notes = []
@@ -80,9 +85,8 @@ class DVMBasicBlock:
         """
         Get all instructions from a basic block.
 
-        :rtype: Return all instructions in the current basic block
+        :returns: Return all instructions in the current basic block
         """
-        tmp_ins = []
         idx = 0
         for i in self.method.get_instructions():
             if self.start <= idx < self.end:
@@ -93,33 +97,58 @@ class DVMBasicBlock:
         return self.nb_instructions
 
     def get_method(self):
+        """
+        Returns the originiating method
+
+        :return: the method
+        :rtype: androguard.core.bytecodes.dvm.EncodedMethod
+        """
         return self.method
 
     def get_name(self):
-        return "{}-BB@0x{:x}".format(self.method.get_name(), self.start)
+        return self.name
 
     def get_start(self):
+        """
+        Get the starting offset of this basic block
+
+        :return: starting offset
+        :rtype: int
+        """
         return self.start
 
     def get_end(self):
+        """
+        Get the end offset of this basic block
+
+        :return: end offset
+        :rtype: int
+        """
         return self.end
 
     def get_last(self):
-        return self.get_instructions()[-1]
+        """
+        Get the last instruction in the basic block
+
+        :return: androguard.core.bytecodes.dvm.Instruction
+        """
+        return list(self.get_instructions())[-1]
 
     def get_next(self):
         """
-            Get next basic blocks
+        Get next basic blocks
 
-            :rtype: a list of the next basic blocks
+        :returns: a list of the next basic blocks
+        :rtype: DVMBasicBlock
         """
         return self.childs
 
     def get_prev(self):
         """
-            Get previous basic blocks
+        Get previous basic blocks
 
-            :rtype: a list of the previous basic blocks
+        :returns: a list of the previous basic blocks
+        :rtype: DVMBasicBlock
         """
         return self.fathers
 
@@ -162,11 +191,11 @@ class DVMBasicBlock:
 
     def get_special_ins(self, idx):
         """
-            Return the associated instruction to a specific instruction (for example a packed/sparse switch)
+        Return the associated instruction to a specific instruction (for example a packed/sparse switch)
 
-            :param idx: the index of the instruction
+        :param idx: the index of the instruction
 
-            :rtype: None or an Instruction
+        :rtype: None or an Instruction
         """
         if idx in self.special_ins:
             return self.special_ins[idx]
@@ -180,19 +209,27 @@ class DVMBasicBlock:
         self.exception_analysis = exception_analysis
 
     def show(self):
-        print(self.get_name(), self.get_start(), self.get_end())
+        print("{}: {:04x} - {:04x}".format(self.get_name(), self.get_start(), self.get_end()))
+        for note in self.get_notes():
+            print(note)
+        print('=' * 20)
 
 
 class BasicBlocks:
     """
-        This class represents all basic blocks of a method
-    """
+    This class represents all basic blocks of a method.
 
-    def __init__(self, _vm):
-        self.__vm = _vm
+    It is a collection of many :class:`DVMBasicBlock`.
+    """
+    def __init__(self):
         self.bb = []
 
     def push(self, bb):
+        """
+        Adds another basic block to the collection
+
+        :param DVBMBasicBlock bb: the DVMBasicBlock to add
+        """
         self.bb.append(bb)
 
     def pop(self, idx):
@@ -204,21 +241,35 @@ class BasicBlocks:
                 return i
         return None
 
-    def get(self):
+    def __len__(self):
+        return len(self.bb)
+
+    def __iter__(self):
         """
-            :rtype: return each basic block (:class:`DVMBasicBlock` object)
+        :returns: yields each basic block (:class:`DVMBasicBlock` object)
+        :rtype: Iterator[DVMBasicBlock]
         """
-        for i in self.bb:
-            yield i
+        yield from self.bb
+
+    def __getitem__(self, item):
+        """
+        Get the basic block at the index
+
+        :param item: index
+        :return: The basic block
+        :rtype: DVMBasicBlock
+        """
+        return self.bb[item]
 
     def gets(self):
         """
-            :rtype: a list of basic blocks (:class:`DVMBasicBlock` objects)
+        :returns: a list of basic blocks (:class:`DVMBasicBlock` objects)
         """
         return self.bb
 
-    def get_basic_block_pos(self, idx):
-        return self.bb[idx]
+    # Alias for legacy programs
+    get = __iter__
+    get_basic_block_pos = __getitem__
 
 
 class ExceptionAnalysis:
@@ -252,8 +303,7 @@ class ExceptionAnalysis:
 
 
 class Exceptions:
-    def __init__(self, _vm):
-        self.__vm = _vm
+    def __init__(self):
         self.exceptions = []
 
     def add(self, exceptions, basic_blocks):
@@ -279,44 +329,96 @@ class Exceptions:
 
 
 class MethodAnalysis:
-    def __init__(self, vm, method):
-        """
-        This class analyses in details a method of a class/dex file
-        It is a wrapper around a :class:`EncodedMethod` and enhances it
-        by using multiple :class:`BasicBlock`.
+    """
+    This class analyses in details a method of a class/dex file
+    It is a wrapper around a :class:`EncodedMethod` and enhances it
+    by using multiple :class:`DVMBasicBlock` encapsulated in a :class:`BasicBlocks` object.
 
-        :type vm: a :class:`DalvikVMFormat` object
-        :type method: a :class:`EncodedMethod` object
-        """
+    :type vm: a :class:`DalvikVMFormat` object
+    :type method: a :class:`EncodedMethod` object
+    """
+    def __init__(self, vm, method):
         self.__vm = vm
         self.method = method
 
-        self.basic_blocks = BasicBlocks(self.__vm)
-        self.exceptions = Exceptions(self.__vm)
+        self.basic_blocks = BasicBlocks()
+        self.exceptions = Exceptions()
 
-        self.code = self.method.get_code()
+        self.xrefto = set()
+        self.xreffrom = set()
+
+        self.xrefread = set()
+        self.xrefwrite = set()
+
+        self.xrefnewinstance = set()
+        self.xrefconstclass = set()
+
+        # Reserved for further use
+        self.apilist = None
+
+        if vm is None or isinstance(method, ExternalMethod):
+            # Support external methods here
+            # external methods usually dont have a VM associated
+            self.code = None
+        else:
+            self.code = self.method.get_code()
+
         if self.code:
             self._create_basic_block()
 
+    @property
+    def name(self):
+        """Returns the name of this method"""
+        return self.method.get_name()
+
+    @property
+    def descriptor(self):
+        """Returns the type descriptor for this method"""
+        return self.method.get_descriptor()
+
+    @property
+    def access(self):
+        """Returns the access flags to the method as a string"""
+        return self.method.get_access_flags_string()
+
+    @property
+    def class_name(self):
+        """Returns the name of the class of this method"""
+        return self.method.class_name
+
+    @property
+    def full_name(self):
+        """Returns classname + name + descriptor, separated by spaces (no access flags)"""
+        return self.method.full_name
+
+    def get_class_name(self):
+        """Return the class name of the method"""
+        return self.class_name
+
+    def get_access_flags_string(self):
+        """Returns the concatenated access flags string"""
+        return self.access
+
+    def get_descriptor(self):
+        return self.descriptor
+
     def _create_basic_block(self):
+        """
+        Internal Method to create the basic block structure
+        Parses all instructions and exceptions.
+        """
         current_basic = DVMBasicBlock(0, self.__vm, self.method, self.basic_blocks)
         self.basic_blocks.push(current_basic)
 
-        bc = self.code.get_bc()
         l = []
-        h = {}
-        idx = 0
+        h = dict()
 
-        log.debug("Parsing instructions")
-        for i in bc.get_instructions():
-            for j in BasicOPCODES:
-                if j.match(i.get_name()) is not None:
-                    v = dvm.determineNext(i, idx, self.method)
-                    h[idx] = v
-                    l.extend(v)
-                    break
-
-            idx += i.get_length()
+        log.debug("Parsing instructions for method at @0x{:08x}".format(self.method.get_code_off()))
+        for idx, ins in self.method.get_instructions_idx():
+            if ins.get_op_value() in BasicOPCODES:
+                v = dvm.determineNext(ins, idx, self.method)
+                h[idx] = v
+                l.extend(v)
 
         log.debug("Parsing exceptions")
         excepts = dvm.determineException(self.__vm, self.method)
@@ -325,29 +427,25 @@ class MethodAnalysis:
             for handler in i[2:]:
                 l.append(handler[1])
 
-        log.debug("Creating basic blocks in %s" % self.method)
-        idx = 0
-        for i in bc.get_instructions():
+        log.debug("Creating basic blocks")
+        for idx, ins in self.method.get_instructions_idx():
             # index is a destination
             if idx in l:
                 if current_basic.get_nb_instructions() != 0:
                     current_basic = DVMBasicBlock(current_basic.get_end(), self.__vm, self.method, self.basic_blocks)
                     self.basic_blocks.push(current_basic)
 
-            current_basic.push(i)
+            current_basic.push(ins)
 
             # index is a branch instruction
             if idx in h:
                 current_basic = DVMBasicBlock(current_basic.get_end(), self.__vm, self.method, self.basic_blocks)
                 self.basic_blocks.push(current_basic)
 
-            idx += i.get_length()
-
         if current_basic.get_nb_instructions() == 0:
             self.basic_blocks.pop(-1)
 
         log.debug("Settings basic blocks childs")
-
         for i in self.basic_blocks.get():
             try:
                 i.set_childs(h[i.end - i.get_last_length()])
@@ -355,30 +453,200 @@ class MethodAnalysis:
                 i.set_childs([])
 
         log.debug("Creating exceptions")
-
-        # Create exceptions
         self.exceptions.add(excepts, self.basic_blocks)
 
         for i in self.basic_blocks.get():
             # setup exception by basic block
             i.set_exception_analysis(self.exceptions.get_exception(i.start, i.end - 1))
 
+    def add_xref_read(self, classobj, fieldobj, offset):
+        """
+        :param ClassAnalysis classobj:
+        :param FieldAnalysis fieldobj:
+        :param int offset: offset in the bytecode
+        """
+        self.xrefread.add((classobj, fieldobj, offset))
+
+    def add_xref_write(self, classobj, fieldobj, offset):
+        """
+        :param ClassAnalysis classobj:
+        :param FieldAnalysis fieldobj:
+        :param int offset: offset in the bytecode
+        """
+        self.xrefwrite.add((classobj, fieldobj, offset))
+
+    def get_xref_read(self):
+        """
+        Returns a list of xrefs where a field is read by this method.
+
+        The list contains tuples of the originating class and methods,
+        where the class is represented as a :class:`ClassAnalysis`,
+        while the Field is a :class:`FieldAnalysis`.
+        """
+        return self.xrefread
+
+    def get_xref_write(self):
+        """
+        Returns a list of xrefs where a field is written to by this method.
+
+        The list contains tuples of the originating class and methods,
+        where the class is represented as a :class:`ClassAnalysis`,
+        while the Field is a :class:`FieldAnalysis`.
+        """
+        return self.xrefwrite
+
+    def add_xref_to(self, classobj, methodobj, offset):
+        """
+        Add a crossreference to another method
+        (this method calls another method)
+
+        :param classobj: :class:`~ClassAnalysis`
+        :param methodobj:  :class:`~androguard.core.bytecodes.dvm.EncodedMethod`
+        :param offset: integer where in the method the call happens
+        """
+        self.xrefto.add((classobj, methodobj, offset))
+
+    def add_xref_from(self, classobj, methodobj, offset):
+        """
+        Add a crossrefernece from another method
+        (this method is called by another method)
+
+        :param classobj: :class:`~ClassAnalysis`
+        :param methodobj:  :class:`~androguard.core.bytecodes.dvm.EncodedMethod`
+        :param offset: integer where in the method the call happens
+        """
+        self.xreffrom.add((classobj, methodobj, offset))
+
+    def get_xref_from(self):
+        """
+        Returns a list of tuples containing the class, method and offset of
+        the call, from where this object was called.
+
+        The list of tuples has the form:
+        (:class:`~ClassAnalysis`,
+        :class:`~androguard.core.bytecodes.dvm.EncodedMethod` or
+        :class:`~ExternalMethod`, :class:`int`)
+        """
+        return self.xreffrom
+
+    def get_xref_to(self):
+        """
+        Returns a list of tuples containing the class, method and offset of
+        the call, which are called by this method.
+
+        The list of tuples has the form:
+        (:class:`~ClassAnalysis`,
+        :class:`~androguard.core.bytecodes.dvm.EncodedMethod` or
+        :class:`~ExternalMethod`, :class:`int`)
+        """
+        return self.xrefto
+
+    def add_xref_new_instance(self, classobj, offset):
+        """
+        Add a crossreference to another class that is
+        instanced within this method.
+
+        :param classobj: :class:`~ClassAnalysis`
+        :param offset: integer where in the method the instantiation happens
+        """
+        self.xrefnewinstance.add((classobj, offset))
+
+    def get_xref_new_instance(self):
+        """
+        Returns a list of tuples containing the class and offset of
+        the creation of a new instance of a class by this method.
+
+        The list of tuples has the form:
+        (:class:`~ClassAnalysis`,
+        :class:`int`)
+        """
+        return self.xrefnewinstance
+
+    def add_xref_const_class(self, classobj, offset):
+        """
+        Add a crossreference to another classtype.
+
+        :param classobj: :class:`~ClassAnalysis`
+        :param offset: integer where in the method the classtype is referenced
+        """
+        self.xrefconstclass.add((classobj, offset))
+
+    def get_xref_const_class(self):
+        """
+        Returns a list of tuples containing the class and offset of
+        the references to another classtype by this method.
+
+        The list of tuples has the form:
+        (:class:`~ClassAnalysis`,
+        :class:`int`)
+        """
+        return self.xrefconstclass
+
+    def is_external(self):
+        """
+        Returns True if the underlying method is external
+
+        :rtype: boolean
+        """
+        return isinstance(self.method, ExternalMethod)
+
+    def is_android_api(self):
+        """
+        Returns True if the method seems to be an Android API method.
+
+        This method might be not very precise unless an list of known API methods
+        is given.
+
+        :return: boolean
+        """
+        if not self.is_external():
+            # Method must be external to be an API
+            return False
+
+        # Packages found at https://developer.android.com/reference/packages.html
+        api_candidates = ["Landroid/", "Lcom/android/internal/util", "Ldalvik/", "Ljava/", "Ljavax/", "Lorg/apache/",
+                          "Lorg/json/", "Lorg/w3c/dom/", "Lorg/xml/sax", "Lorg/xmlpull/v1/", "Ljunit/"]
+
+        if self.apilist:
+            # FIXME: This will not work... need to introduce a name for lookup (like EncodedMethod.__str__ but without
+            # the offset! Such a name is also needed for the lookup in permissions
+            return self.method.get_name() in self.apilist
+        else:
+            for candidate in api_candidates:
+                if self.method.get_class_name().startswith(candidate):
+                    return True
+
+        return False
+
     def get_basic_blocks(self):
         """
-            :rtype: a :class:`BasicBlocks` object
+        Returns the :class:`BasicBlocks` generated for this method.
+        The :class:`BasicBlocks` can be used to get a control flow graph (CFG) of the method.
+
+        :rtype: a :class:`BasicBlocks` object
         """
         return self.basic_blocks
 
     def get_length(self):
         """
-            :rtype: an integer which is the length of the code
+        :returns: an integer which is the length of the code
+        :rtype: int
         """
         return self.code.get_length() if self.code else 0
 
     def get_vm(self):
+        """
+        :rtype: androguard.core.bytecodes.dvm.DalvikVMFormat
+        :return:
+        """
         return self.__vm
 
     def get_method(self):
+        """
+
+        :rtype: androguard.core.bytecodes.dvm.EncodedMethod
+        :return:
+        """
         return self.method
 
     def show(self):
@@ -403,38 +671,64 @@ class MethodAnalysis:
               self.method.get_access_flags_string(),
               self.method.get_name(),
               ", ".join(args), ret))
-        bytecode.PrettyShow(self, self.basic_blocks.gets(), self.method.notes)
+        bytecode.PrettyShow(self.basic_blocks.gets(), self.method.notes)
+
+    def show_xrefs(self):
+        data = "XREFto for %s\n" % self.method
+        for ref_class, ref_method, offset in self.xrefto:
+            data += "in\n"
+            data += "{}:{} @0x{:x}\n".format(ref_class.get_vm_class().get_name(), ref_method, offset)
+
+        data += "XREFFrom for %s\n" % self.method
+        for ref_class, ref_method, offset in self.xreffrom:
+            data += "in\n"
+            data += "{}:{} @0x{:x}\n".format(ref_class.get_vm_class().get_name(), ref_method, offset)
+
+        return data
 
     def __repr__(self):
         return "<analysis.MethodAnalysis {}>".format(self.method)
 
 
 class StringAnalysis:
+    """
+    StringAnalysis contains the XREFs of a string.
+
+    As Strings are only used as a source, they only contain
+    the XREF_FROM set, i.e. where the string is used.
+
+    This Array stores the information in which method the String is used.
+    """
     def __init__(self, value):
         """
-        StringAnalysis contains the XREFs of a string.
 
-        As Strings are only used as a source, they only contain
-        the XREF_FROM set, i.e. where the string is used.
-
-        This Array stores the information in which method the String is used.
+        :param str value: the original string value
         """
         self.value = value
         self.orig_value = value
         self.xreffrom = set()
 
-    def AddXrefFrom(self, classobj, methodobj):
-        self.xreffrom.add((classobj, methodobj))
+    def add_xref_from(self, classobj, methodobj, off):
+        """
+        Adds a xref from the given method to this string
 
-    def get_xref_from(self):
+        :param ClassAnalysis classobj:
+        :param MethodAnalysis methodobj:
+        :param int off: offset in the bytecode of the call
+        """
+        self.xreffrom.add((classobj, methodobj, off))
+
+    def get_xref_from(self, withoffset=False):
         """
         Returns a list of xrefs accessing the String.
 
         The list contains tuples of the originating class and methods,
         where the class is represented as a :class:`ClassAnalysis`,
-        while the method is a :class:`androguard.core.bytecodes.dvm.EncodedMethod`.
+        while the method is a :class:`MethodAnalysis`.
         """
-        return self.xreffrom
+        if withoffset:
+            return self.xreffrom
+        return set(map(itemgetter(slice(0, 2)), self.xreffrom))
 
     def set_value(self, value):
         """
@@ -483,167 +777,18 @@ class StringAnalysis:
         return "<analysis.StringAnalysis {}>".format(s)
 
 
-class MethodClassAnalysis:
-    def __init__(self, method):
-        """
-        MethodClassAnalysis contains the XREFs for a given method.
+class FieldAnalysis:
+    """
+    FieldAnalysis contains the XREFs for a class field.
 
-        Both referneces to other methods (XREF_TO) as well as methods calling
-        this method (XREF_FROM) are saved.
+    Instead of using XREF_FROM/XREF_TO, this object has methods for READ and
+    WRITE access to the field.
 
-        :param androguard.core.bytecodes.dvm.EncodedMethod method: the DVM Method object
-        """
-        self.method = method
-        self.xrefto = set()
-        self.xreffrom = set()
+    That means, that it will show you, where the field is read or written.
 
-        # Reserved for further use
-        self.apilist = None
-
-    @property
-    def name(self):
-        """Returns the name of this method"""
-        return self.method.get_name()
-
-    @property
-    def descriptor(self):
-        """Returns the type descriptor for this method"""
-        return self.method.get_descriptor()
-
-    @property
-    def access(self):
-        """Returns the access flags to the method as a string"""
-        return self.method.get_access_flags_string()
-
-    @property
-    def class_name(self):
-        """Returns the name of the class of this method"""
-        return self.method.class_name
-
-    @property
-    def full_name(self):
-        """Returns classname + name + descriptor, separated by spaces (no access flags)"""
-        return self.method.full_name
-
-    def AddXrefTo(self, classobj, methodobj, offset):
-        """
-        Add a crossreference to another method
-        (this method calls another method)
-
-        :param classobj: :class:`~ClassAnalysis`
-        :param methodobj:  :class:`~androguard.core.bytecodes.dvm.EncodedMethod`
-        :param offset: integer where in the method the call happens
-        """
-        self.xrefto.add((classobj, methodobj, offset))
-
-    def AddXrefFrom(self, classobj, methodobj, offset):
-        """
-        Add a crossrefernece from another method
-        (this method is called by another method)
-
-        :param classobj: :class:`~ClassAnalysis`
-        :param methodobj:  :class:`~androguard.core.bytecodes.dvm.EncodedMethod`
-        :param offset: integer where in the method the call happens
-        """
-        self.xreffrom.add((classobj, methodobj, offset))
-
-    def get_xref_from(self):
-        """
-        Returns a list of tuples containing the class, method and offset of
-        the call, from where this object was called.
-
-        The list of tuples has the form:
-        (:class:`~ClassAnalysis`,
-        :class:`~androguard.core.bytecodes.dvm.EncodedMethod` or
-        :class:`~ExternalMethod`, :class:`int`)
-        """
-        return self.xreffrom
-
-    def get_xref_to(self):
-        """
-        Returns a list of tuples containing the class, method and offset of
-        the call, which are called by this method.
-
-        The list of tuples has the form:
-        (:class:`~ClassAnalysis`,
-        :class:`~androguard.core.bytecodes.dvm.EncodedMethod` or
-        :class:`~ExternalMethod`, :class:`int`)
-        """
-        return self.xrefto
-
-    def is_external(self):
-        """
-        Returns True if the underlying method is external
-
-        :rtype: boolean
-        """
-        return isinstance(self.method, ExternalMethod)
-
-    def is_android_api(self):
-        """
-        Returns True if the method seems to be an Android API method.
-
-        This method might be not very precise unless an list of known API methods
-        is given.
-
-        :return: boolean
-        """
-        if not self.is_external():
-            # Method must be external to be an API
-            return False
-
-        # Packages found at https://developer.android.com/reference/packages.html
-        api_candidates = ["Landroid/", "Lcom/android/internal/util", "Ldalvik/", "Ljava/", "Ljavax/", "Lorg/apache/",
-                          "Lorg/json/", "Lorg/w3c/dom/", "Lorg/xml/sax", "Lorg/xmlpull/v1/", "Ljunit/"]
-
-        if self.apilist:
-            # FIXME: This will not work... need to introduce a name for lookup (like EncodedMethod.__str__ but without
-            # the offset! Such a name is also needed for the lookup in permissions
-            return self.method.get_name() in self.apilist
-        else:
-            for candidate in api_candidates:
-                if self.method.get_class_name().startswith(candidate):
-                    return True
-
-        return False
-
-    def get_method(self):
-        """
-        Return the `EncodedMethod` object that relates to this object
-        :return: `dvm.EncodedMethod`
-        """
-        return self.method
-
-    def __str__(self):
-        data = "XREFto for %s\n" % self.method
-        for ref_class, ref_method, offset in self.xrefto:
-            data += "in\n"
-            data += "{}:{} @0x{:x}\n".format(ref_class.get_vm_class().get_name(), ref_method, offset)
-
-        data += "XREFFrom for %s\n" % self.method
-        for ref_class, ref_method, offset in self.xreffrom:
-            data += "in\n"
-            data += "{}:{} @0x{:x}\n".format(ref_class.get_vm_class().get_name(), ref_method, offset)
-
-        return data
-
-    def __repr__(self):
-        return "<analysis.MethodClassAnalysis {}{}>".format(self.method,
-               " EXTERNAL" if isinstance(self.method, ExternalMethod) else "")
-
-
-class FieldClassAnalysis:
+    :param androguard.core.bytecodes.dvm.EncodedField field: `dvm.EncodedField`
+    """
     def __init__(self, field):
-        """
-        FieldClassAnalysis contains the XREFs for a class field.
-
-        Instead of using XREF_FROM/XREF_TO, this object has methods for READ and
-        WRITE access to the field.
-
-        That means, that it will show you, where the field is read or written.
-
-        :param androguard.core.bytecodes.dvm.EncodedField field: `dvm.EncodedField`
-        """
         self.field = field
         self.xrefread = set()
         self.xrefwrite = set()
@@ -652,93 +797,97 @@ class FieldClassAnalysis:
     def name(self):
         return self.field.get_name()
 
-    def AddXrefRead(self, classobj, methodobj):
-        self.xrefread.add((classobj, methodobj))
+    def add_xref_read(self, classobj, methodobj, offset):
+        """
+        :param ClassAnalysis classobj:
+        :param MethodAnalysis methodobj:
+        :param int offset: offset in the bytecode
+        """
+        self.xrefread.add((classobj, methodobj, offset))
 
-    def AddXrefWrite(self, classobj, methodobj):
-        self.xrefwrite.add((classobj, methodobj))
+    def add_xref_write(self, classobj, methodobj, offset):
+        """
+        :param ClassAnalysis classobj:
+        :param MethodAnalysis methodobj:
+        :param int offset: offset in the bytecode
+        """
+        self.xrefwrite.add((classobj, methodobj, offset))
 
-    def get_xref_read(self):
+    def get_xref_read(self, withoffset=False):
         """
         Returns a list of xrefs where the field is read.
 
         The list contains tuples of the originating class and methods,
         where the class is represented as a :class:`ClassAnalysis`,
-        while the method is a :class:`androguard.core.bytecodes.dvm.EncodedMethod`.
-        """
-        return self.xrefread
+        while the method is a :class:`MethodAnalysis`.
 
-    def get_xref_write(self):
+        :param bool withoffset: return the xrefs including the offset
+        """
+        if withoffset:
+            return self.xrefread
+        # Legacy option, might be removed in the future
+        return set(map(itemgetter(slice(0, 2)), self.xrefread))
+
+    def get_xref_write(self, withoffset=False):
         """
         Returns a list of xrefs where the field is written to.
 
         The list contains tuples of the originating class and methods,
         where the class is represented as a :class:`ClassAnalysis`,
-        while the method is a :class:`androguard.core.bytecodes.dvm.EncodedMethod`.
+        while the method is a :class:`MethodAnalysis`.
+
+        :param bool withoffset: return the xrefs including the offset
         """
-        return self.xrefwrite
+        if withoffset:
+            return self.xrefwrite
+        # Legacy option, might be removed in the future
+        return set(map(itemgetter(slice(0, 2)), self.xrefwrite))
 
     def get_field(self):
+        """
+        Returns the actual field object
+
+        :rtype: androguard.core.bytecodes.dvm.EncodedField
+        """
         return self.field
 
     def __str__(self):
         data = "XREFRead for %s\n" % self.field
-        for ref_class, ref_method in self.xrefread:
+        for ref_class, ref_method, off in self.xrefread:
             data += "in\n"
-            data += "{}:{}\n".format(ref_class.get_vm_class().get_name(), ref_method)
+            data += "{}:{} @{}\n".format(ref_class.get_vm_class().get_name(), ref_method, off)
 
         data += "XREFWrite for %s\n" % self.field
-        for ref_class, ref_method in self.xrefwrite:
+        for ref_class, ref_method, off in self.xrefwrite:
             data += "in\n"
-            data += "{}:{}\n".format(ref_class.get_vm_class().get_name(), ref_method)
+            data += "{}:{} @{}\n".format(ref_class.get_vm_class().get_name(), ref_method, off)
 
         return data
 
     def __repr__(self):
-        return "<analysis.FieldClassAnalysis {}->{}>".format(self.field.class_name, self.field.name)
+        return "<analysis.FieldAnalysis {}->{}>".format(self.field.class_name, self.field.name)
 
 
 class ExternalClass:
-    def __init__(self, name):
-        """
-        The ExternalClass is used for all classes that are not defined in the
-        DEX file, thus are external classes.
+    """
+    The ExternalClass is used for all classes that are not defined in the
+    DEX file, thus are external classes.
 
-        :param name: Name of the external class
-        """
+    :param name: Name of the external class
+    """
+    def __init__(self, name):
         self.name = name
-        self.methods = {}
+        self.methods = []
 
     def get_methods(self):
         """
         Return the stored methods for this external class
         :return:
         """
-        return self.methods.values()
+        return self.methods
 
-    def GetMethod(self, name, descriptor):
-        """
-        .. deprecated:: 3.1.0
-            Use :meth:`get_method` instead.
-
-        """
-        warnings.warn("deprecated, use get_method instead. This function might be removed in a later release!", DeprecationWarning)
-        return self.get_method(name, descriptor)
-
-    def get_method(self, name, descriptor):
-        """
-        Get the method by name and descriptor,
-        or create a new one if the requested method does not exists.
-
-        :param name: method name
-        :param descriptor: method descriptor, for example `'(I)V'`
-        :return: :class:`ExternalMethod`
-        """
-        key = name + str(descriptor)
-        if key not in self.methods:
-            self.methods[key] = ExternalMethod(self.name, name, descriptor)
-
-        return self.methods[key]
+    def add_method(self, method):
+        self.methods.append(method)
 
     def get_name(self):
         """
@@ -751,6 +900,21 @@ class ExternalClass:
 
 
 class ExternalMethod:
+    """
+    ExternalMethod is a stub class for methods that are not part of the current Analysis.
+    There are two possibilities for this:
+
+    1) The method is defined inside another DEX file which was not loaded into the Analysis
+    2) The method is an API method, hence it is defined in the Android system
+
+    External methods should have a similar API to :class:`~androguard.core.bytecodes.dvm.EncodedMethod`
+    but obviously they have no code attached.
+    The only known information about such methods are the class name, the method name and its descriptor.
+
+    :param str class_name: name of the class
+    :param str name: name of the method
+    :param List[str] descriptor: descriptor string
+    """
     def __init__(self, class_name, name, descriptor):
         self.class_name = class_name
         self.name = name
@@ -763,57 +927,82 @@ class ExternalMethod:
         return self.class_name
 
     def get_descriptor(self):
-        return ''.join(self.descriptor)
+        return mutf8.MUTF8String.join(self.descriptor)
 
     @property
     def full_name(self):
         """Returns classname + name + descriptor, separated by spaces (no access flags)"""
-        return self.class_name + " " + self.name + " " + self.get_descriptor()
+        return self.class_name + " " + self.name + " " + str(self.get_descriptor())
 
     @property
     def permission_api_name(self):
         """Returns a name which can be used to look up in the permission maps"""
-        return self.class_name + "-" + self.name + "-" + self.get_descriptor()
+        return self.class_name + "-" + self.name + "-" + str(self.get_descriptor())
 
     def get_access_flags_string(self):
+        """
+        Returns the access flags string.
+
+        Right now, this is always an empty strings, as we can not say what
+        kind of access flags an external method might have.
+        """
         # TODO can we assume that external methods are always public?
         # they can also be static...
         # or constructor...
+        # or they might be inherited and have all kinds of access flags...
         return ""
 
     def __str__(self):
-        return "{}->{}{}".format(self.class_name, self.name, ''.join(self.descriptor))
+        return "{}->{}{}".format(self.class_name.__str__(), self.name.__str__(), str(self.get_descriptor()))
 
     def __repr__(self):
         return "<analysis.ExternalMethod {}>".format(self.__str__())
 
 
 class ClassAnalysis:
+    """
+    ClassAnalysis contains the XREFs from a given Class.
+    It is also used to wrap :class:`~androguard.core.bytecode.dvm.ClassDefItem`, which
+    contain the actual class content like bytecode.
+
+    Also external classes will generate xrefs, obviously only XREF_FROM are
+    shown for external classes.
+
+    :param classobj: class:`~androguard.core.bytecode.dvm.ClassDefItem` or :class:`ExternalClass`
+    """
+
     def __init__(self, classobj):
-        """
-        ClassAnalysis contains the XREFs from a given Class.
-        It is also used to wrap :class:`~androguard.core.bytecode.dvm.ClassDefItem`, which
-        contain the actual class content like bytecode.
-
-        Also external classes will generate xrefs, obviously only XREF_FROM are
-        shown for external classes.
-
-        :param classobj: class:`~androguard.core.bytecode.dvm.ClassDefItem` or :class:`ExternalClass`
-        """
-
         # Automatically decide if the class is external or not
         self.external = isinstance(classobj, ExternalClass)
 
         self.orig_class = classobj
-        self._inherits_methods = {}
-        self._methods = {}
-        self._fields = {}
+
+        # Contains EncodedMethod/ExternalMethod -> MethodAnalysis
+        self._methods = dict()
+
+        # Contains EncodedField -> FieldAnalysis
+        self._fields = dict()
 
         self.xrefto = collections.defaultdict(set)
         self.xreffrom = collections.defaultdict(set)
 
+        self.xrefnewinstance = set()
+        self.xrefconstclass = set()
+
         # Reserved for further use
         self.apilist = None
+
+    def add_method(self, method_analysis):
+        """
+        Add the given method to this analyis.
+        usually only called during Analysis.add and Analysis._resolve_method
+
+        :param MethodAnalysis method_analysis:
+        """
+        self._methods[method_analysis.get_method()] = method_analysis
+        if self.external:
+            # Propagate ExternalMethod to ExternalClass
+            self.orig_class.add_method(method_analysis.get_method())
 
     @property
     def implements(self):
@@ -852,7 +1041,7 @@ class ClassAnalysis:
 
     def is_external(self):
         """
-        Tests wheather this class is an external class
+        Tests if this class is an external class
 
         :return: True if the Class is external, False otherwise
         """
@@ -888,13 +1077,15 @@ class ClassAnalysis:
 
     def get_methods(self):
         """
-        Return all :class:`MethodClassAnalysis` objects of this class
+        Return all :class:`MethodAnalysis` objects of this class
+
+        :rtype: Iterator[MethodAnalysis]
         """
         return list(self._methods.values())
 
     def get_fields(self):
         """
-        Return all `FieldClassAnalysis` objects of this class
+        Return all `FieldAnalysis` objects of this class
         """
         return self._fields.values()
 
@@ -906,95 +1097,104 @@ class ClassAnalysis:
 
     def get_method_analysis(self, method):
         """
-        Return the MethodClassAnalysis object for a given EncodedMethod
+        Return the MethodAnalysis object for a given EncodedMethod
 
         :param method: :class:`EncodedMethod`
-        :return: :class:`MethodClassAnalysis`
+        :return: :class:`MethodAnalysis`
+        :rtype: MethodAnalysis
         """
         return self._methods.get(method)
 
     def get_field_analysis(self, field):
         return self._fields.get(field)
 
-    def get_fake_method(self, name, descriptor):
-        """
-        Search for the given method name and descriptor
-        and return a fake (ExternalMethod) if required.
-
-        :param name: name of the method
-        :param descriptor: descriptor of the method, for example `'(I I I)V'`
-        :return: :class:`ExternalMethod`
-        """
-        if self.external:
-            # An external class can only generate the methods on demand
-            return self.orig_class.get_method(name, descriptor)
-
-        # We are searching an unknown method in this class
-        # It could be something that the class herits
-        key = name + str(descriptor)
-        if key not in self._inherits_methods:
-            self._inherits_methods[key] = ExternalMethod(self.orig_class.get_name(), name, descriptor)
-        return self._inherits_methods[key]
-
-    def AddFXrefRead(self, method, classobj, field):
+    def add_field_xref_read(self, method, classobj, field, off):
         """
         Add a Field Read to this class
 
-        :param method:
-        :param classobj:
-        :param field:
+        :param MethodAnalysis method:
+        :param ClassAnalysis classobj:
+        :param androguard.code.bytecodes.dvm.EncodedField field:
+        :param int off:
         :return:
         """
         if field not in self._fields:
-            self._fields[field] = FieldClassAnalysis(field)
-        self._fields[field].AddXrefRead(classobj, method)
+            self._fields[field] = FieldAnalysis(field)
+        self._fields[field].add_xref_read(classobj, method, off)
 
-    def AddFXrefWrite(self, method, classobj, field):
+    def add_field_xref_write(self, method, classobj, field, off):
         """
-        Add a Field Write to this class
+        Add a Field Write to this class in a given method
 
-        :param method:
-        :param classobj:
-        :param field:
+        :param MethodAnalysis method:
+        :param ClassAnalysis classobj:
+        :param androguard.core.bytecodes.dvm.EncodedField field:
+        :param int off:
         :return:
         """
         if field not in self._fields:
-            self._fields[field] = FieldClassAnalysis(field)
-        self._fields[field].AddXrefWrite(classobj, method)
+            self._fields[field] = FieldAnalysis(field)
+        self._fields[field].add_xref_write(classobj, method, off)
 
-    def AddMXrefTo(self, method1, classobj, method2, offset):
-        if method1 not in self._methods:
-            self._methods[method1] = MethodClassAnalysis(method1)
-        self._methods[method1].AddXrefTo(classobj, method2, offset)
+    def add_method_xref_to(self, method1, classobj, method2, offset):
+        """
 
-    def AddMXrefFrom(self, method1, classobj, method2, offset):
-        if method1 not in self._methods:
-            self._methods[method1] = MethodClassAnalysis(method1)
-        self._methods[method1].AddXrefFrom(classobj, method2, offset)
+        :param MethodAnalysis method1: the calling method
+        :param ClassAnalysis classobj: the calling class
+        :param MethodAnalysis method2: the called method
+        :param int offset: offset in the bytecode of calling method
+        """
 
-    def AddXrefTo(self, ref_kind, classobj, methodobj, offset):
+        # FIXME: Not entirely sure why this can happen but usually a multidex issue:
+        # The given method was not added before...
+        if method1.get_method() not in self._methods:
+            self.add_method(method1)
+
+        self._methods[method1.get_method()].add_xref_to(classobj, method2, offset)
+
+    def add_method_xref_from(self, method1, classobj, method2, offset):
+        """
+
+        :param MethodAnalysis method1:
+        :param ClassAnalysis classobj:
+        :param MethodAnalysis method2:
+        :param int offset:
+        """
+        # FIXME: Not entirely sure why this can happen but usually a multidex issue:
+        # The given method was not added before...
+        if method1.get_method() not in self._methods:
+            self.add_method(method1)
+
+        self._methods[method1.get_method()].add_xref_from(classobj, method2, offset)
+
+    def add_xref_to(self, ref_kind, classobj, methodobj, offset):
         """
         Creates a crossreference to another class.
         XrefTo means, that the current class calls another class.
         The current class should also be contained in the another class' XrefFrom list.
 
+        .. warning::
+            The implementation of this specific method might not be what you expect!
+            the parameter :code:`methodobj` is the source method and not the destination
+            in the case that :code:`ref_kind` is const-class or new-instance!
+
         :param REF_TYPE ref_kind: type of call
-        :param classobj: :class:`ClassAnalysis` object to link
-        :param methodobj:
-        :param offset: Offset in the Methods Bytecode, where the call happens
+        :param ClassAnalysis classobj: :class:`ClassAnalysis` object to link
+        :param MethodAnalysis methodobj:
+        :param int offset: Offset in the Methods Bytecode, where the call happens
         :return:
         """
         self.xrefto[classobj].add((ref_kind, methodobj, offset))
 
-    def AddXrefFrom(self, ref_kind, classobj, methodobj, offset):
+    def add_xref_from(self, ref_kind, classobj, methodobj, offset):
         """
         Creates a crossreference from this class.
         XrefFrom means, that the current class is called by another class.
 
         :param REF_TYPE ref_kind: type of call
-        :param classobj: :class:`ClassAnalysis` object to link
-        :param methodobj:
-        :param offset: Offset in the methods bytecode, where the call happens
+        :param ClassAnalysis classobj: :class:`ClassAnalysis` object to link
+        :param MethodAnalysis methodobj:
+        :param int offset: Offset in the methods bytecode, where the call happens
         :return:
         """
         self.xreffrom[classobj].add((ref_kind, methodobj, offset))
@@ -1008,8 +1208,7 @@ class ClassAnalysis:
 
         The dictionary contains the classes as keys (stored as :class:`ClassAnalysis`)
         and has a tuple as values, where the first item is the ref_kind (which is an Enum of type :class:`REF_TYPE`),
-        the second one is the method in which the class is called (either :class:`ExternalMethod` if external or
-        :class:`androguard.core.bytecodes.dvm.EncodedMethod` if internal)
+        the second one is the method in which the class is called (:class:`MethodAnalysis`)
         and the third the offset in the method where the call is originating.
 
         example::
@@ -1021,6 +1220,8 @@ class ClassAnalysis:
                     for ref_kind, ref_method, ref_offset in refs:
                         print("    in method {} {}".format(ref_kind, ref_method))
 
+
+        :rtype: Iterator[Tuple[REF_TYPE, MethodAnalysis, int]]
         """
         return self.xreffrom
 
@@ -1033,8 +1234,7 @@ class ClassAnalysis:
 
         The dictionary contains the classes as keys (stored as :class:`ClassAnalysis`)
         and has a tuple as values, where the first item is the ref_kind (which is an Enum of type :class:`REF_TYPE`),
-        the second one is the method called (either :class:`ExternalMethod` if external or
-        :class:`androguard.core.bytecodes.dvm.EncodedMethod` if internal)
+        the second one is the method called (:class:`MethodAnalysis`)
         and the third the offset in the method where the call is originating.
 
         example::
@@ -1045,18 +1245,71 @@ class ClassAnalysis:
                     print("  calling class {}".format(calling.name))
                     for ref_kind, ref_method, ref_offset in refs:
                         print("    calling method {} {}".format(ref_kind, ref_method))
+
+        :rtype: Iterator[Tuple[REF_TYPE, MethodAnalysis, int]]
         """
         return self.xrefto
 
+    def add_xref_new_instance(self, methobj, offset):
+        """
+        Add a crossreference to another method that is
+        instancing this class.
+
+        :param classobj: :class:`~MethodAnalysis`
+        :param offset: integer where in the method the instantiation happens
+        """
+        self.xrefnewinstance.add((methobj, offset))
+
+    def get_xref_new_instance(self):
+        """
+        Returns a list of tuples containing the set of methods
+        with offsets that instance this class
+
+
+        The list of tuples has the form:
+        (:class:`~MathodAnalysis`,
+        :class:`int`)
+        """
+        return self.xrefnewinstance
+
+    def add_xref_const_class(self, methobj, offset):
+        """
+        Add a crossreference to a method referencing this classtype.
+
+        :param classobj: :class:`~MethodAnalysis`
+        :param offset: integer where in the method the classtype is referenced
+        """
+        self.xrefconstclass.add((methobj, offset))
+
+    def get_xref_const_class(self):
+        """
+        Returns a list of tuples containing the method and offset
+        referencing this classtype.
+
+        The list of tuples has the form:
+        (:class:`~MethodAnalysis`,
+        :class:`int`)
+        """
+        return self.xrefconstclass
+
     def get_vm_class(self):
+        """
+        Returns the original Dalvik VM class or the external class object.
+
+        :return:
+        :rtype: Union[androguard.core.bytecodes.dvm.ClassDefItem, ExternalClass]
+        """
         return self.orig_class
+
+    # Alias
+    get_class = get_vm_class
 
     def __repr__(self):
         return "<analysis.ClassAnalysis {}{}>".format(self.orig_class.get_name(),
                 " EXTERNAL" if isinstance(self.orig_class, ExternalClass) else "")
 
     def __str__(self):
-        # Print only instanciation from other classes here
+        # Print only instantiation from other classes here
         # TODO also method xref and field xref should be printed?
         data = "XREFto for %s\n" % self.orig_class
         for ref_class in self.xrefto:
@@ -1079,58 +1332,83 @@ class ClassAnalysis:
         return data
 
 
+class MethodClassAnalysis(MethodAnalysis):
+    """
+
+    .. deprecated:: 3.4.0
+
+        Always use MethodAnalysis!
+        This method is just here for compatability
+
+    """
+    def __init__(self, meth):
+        super().__init__(meth.cm.vm, meth)
+
+
 class Analysis:
+    """
+    Analysis Object
+
+    The Analysis contains a lot of information about (multiple) DalvikVMFormat objects
+    Features are for example XREFs between Classes, Methods, Fields and Strings.
+    Yet another part is the creation of BasicBlocks, which is important in the usage of
+    the Androguard Decompiler.
+
+    Multiple DalvikVMFormat Objects can be added using the function :meth:`add`.
+
+    XREFs are created for:
+    * classes (`ClassAnalysis`)
+    * methods (`MethodAnalysis`)
+    * strings (`StringAnalyis`)
+    * fields (`FieldAnalysis`)
+
+    The Analysis should be the only object you are using next to the :class:`~androguard.core.bytecodes.apk.APK`.
+    It encapsulates all the Dalvik related functions into a single place, while you have still the ability to use
+    the functions from :class:`~androguard.core.bytecodes.dvm.DalvikVMFormat` and the related classes.
+
+    :param Optional[androguard.core.bytecodes.dvm.DalvikVMFormat] vm: inital DalvikVMFormat object (default None)
+    """
     def __init__(self, vm=None):
-        """
-        Analysis Object
-
-        The Analysis contains a lot of information about (multiple) DalvikVMFormat objects
-        Features are for example XREFs between Classes, Methods, Fields and Strings.
-
-        Multiple DalvikVMFormat Objects can be added using the function `add`
-
-        XREFs are created for:
-        * classes (`ClassAnalysis`)
-        * methods (`MethodClassAnalysis`)
-        * strings (`StringAnalyis`)
-        * fields (`FieldClassAnalysis`)
-
-        :param vm: inital DalvikVMFormat object (default None)
-        """
-
         # Contains DalvikVMFormat objects
         self.vms = []
         # A dict of {classname: ClassAnalysis}, populated on add(vm)
-        self.classes = {}
+        self.classes = dict()
         # A dict of {string: StringAnalysis}, populated on create_xref()
-        self.strings = {}
+        self.strings = dict()
         # A dict of {EncodedMethod: MethodAnalysis}, populated on add(vm)
-        self.methods = {}
+        self.methods = dict()
+
+        # Used to quickly look up methods
+        self.__method_hashes = dict()
 
         if vm:
             self.add(vm)
 
+        self.__created_xrefs = False
+
     def add(self, vm):
         """
-        Add a DalvikVMFormat to this Analysis
+        Add a DalvikVMFormat to this Analysis.
 
-        :param vm: :class:`dvm.DalvikVMFormat` to add to this Analysis
+        :param androguard.core.bytecodes.dvm.DalvikVMFormat vm: :class:`dvm.DalvikVMFormat` to add to this Analysis
         """
         self.vms.append(vm)
+
+        log.info("Adding DEX file version {}".format(vm.version))
+        # TODO: This step can easily be multithreaded, as there is no dependecy between the objects at this stage
+        tic = time.time()
         for current_class in vm.get_classes():
             self.classes[current_class.get_name()] = ClassAnalysis(current_class)
+            for method in current_class.get_methods():
+                self.methods[method] = MethodAnalysis(vm, method)
 
-        for method in vm.get_methods():
-            self.methods[method] = MethodAnalysis(vm, method)
+                self.classes[current_class.get_name()].add_method(self.methods[method])
 
-    def _get_all_classes(self):
-        """
-        Returns all Class objects of all VMs in this Analysis
-        Used by create_xref().
-        """
-        for vm in self.vms:
-            for current_class in vm.get_classes():
-                yield current_class
+                # Store for faster lookup during create_xrefs
+                m_hash = (current_class.get_name(), method.get_name(), str(method.get_descriptor()))
+                self.__method_hashes[m_hash] = self.methods[method]
+
+        log.info("Reading bytecode took : {:0d}min {:02d}s".format(*divmod(int(time.time() - tic), 60)))
 
     def create_xref(self):
         """
@@ -1139,28 +1417,40 @@ class Analysis:
 
         If you are using multiple DEX files, this function must
         be called when all DEX files are added.
-        If you call the function after every DEX file, the
-        crossreferences might be wrong!
+        If you call the function after every DEX file, it will only work
+        for the first time.
         """
+        if self.__created_xrefs:
+            # TODO on concurrent runs, we probably need to clean up first,
+            # or check that we do not write garbage.
+            log.error("You have requested to run create_xref() twice! "
+                      "This will not work and cause problems! This function will exit right now. "
+                      "If you want to add multiple DEX files, use add() several times and then run create_xref() once.")
+            return
+
+        self.__created_xrefs = True
         log.debug("Creating Crossreferences (XREF)")
         tic = time.time()
 
-        # TODO on concurrent runs, we probably need to clean up first,
-        # or check that we do not write garbage.
-
         # TODO multiprocessing
-        for c in self._get_all_classes():
-            self._create_xref(c)
+        # One reason why multiprocessing is hard to implement is the creation of
+        # the external classes and methods. This must be synchronized, which is now possible as we have a single method!
+        for vm in self.vms:
+            for current_class in vm.get_classes():
+                self._create_xref(current_class)
 
-        log.info("End of creating cross references (XREF)")
-        log.info("run time: {:0d}min {:02d}s".format(*divmod(int(time.time() - tic), 60)))
+        # TODO: After we collected all the information, we should add field and
+        # string xrefs to each MethodAnalysis
+
+        log.info("End of creating cross references (XREF) "
+                 "run time: {:0d}min {:02d}s".format(*divmod(int(time.time() - tic), 60)))
 
     def _create_xref(self, current_class):
         """
         Create the xref for `current_class`
 
         There are four steps involved in getting the xrefs:
-        * Xrefs for classes
+        * Xrefs for class instantiation and static class usage
         *       for method calls
         *       for string usage
         *       for field manipulation
@@ -1173,66 +1463,78 @@ class Analysis:
         """
         cur_cls_name = current_class.get_name()
 
-        log.debug("Creating XREF/DREF for %s" % cur_cls_name)
+        log.debug("Creating XREF/DREF for class at @0x{:08x}".format(current_class.get_class_data_off()))
         for current_method in current_class.get_methods():
-            log.debug("Creating XREF for %s" % current_method)
+            log.debug("Creating XREF for method at @0x{:08x}".format(current_method.get_code_off()))
 
-            off = 0
-            for instruction in current_method.get_instructions():
+            cur_meth = self.get_method(current_method)
+            cur_cls = self.classes[cur_cls_name]
+
+            for off, instruction in current_method.get_instructions_idx():
                 op_value = instruction.get_op_value()
 
                 # 1) check for class calls: const-class (0x1c), new-instance (0x22)
                 if op_value in [0x1c, 0x22]:
                     idx_type = instruction.get_ref_kind()
                     # type_info is the string like 'Ljava/lang/Object;'
-                    type_info = instruction.cm.vm.get_cm_type(idx_type)
+                    type_info = instruction.cm.vm.get_cm_type(idx_type).lstrip(b'[')
+                    if type_info[0] != b'L':
+                        # Need to make sure, that we get class types and not other types
+                        continue
 
-                    # Internal xref related to class manipulation
-                    # FIXME should the xref really only set if the class is in self.classes? If an external class is added later, it will be added too!
-                    # See https://github.com/androguard/androguard/blob/d720ebf2a9c8e2a28484f1c81fdddbc57e04c157/androguard/core/analysis/analysis.py#L806
-                    # Before the check would go for internal classes only!
-                    # FIXME: effectively ignoring calls to itself - do we want that?
-                    if type_info != cur_cls_name:
-                        if type_info not in self.classes:
-                            # Create new external class
-                            self.classes[type_info] = ClassAnalysis(ExternalClass(type_info))
+                    if type_info == cur_cls_name:
+                        # FIXME: effectively ignoring calls to itself - do we want that?
+                        continue
 
-                        cur_cls = self.classes[cur_cls_name]
-                        oth_cls = self.classes[type_info]
+                    if type_info not in self.classes:
+                        # Create new external class
+                        self.classes[type_info] = ClassAnalysis(ExternalClass(type_info))
 
-                        # FIXME: xref_to does not work here! current_method is wrong, as it is not the target!
-                        cur_cls.AddXrefTo(REF_TYPE(op_value), oth_cls, current_method, off)
-                        oth_cls.AddXrefFrom(REF_TYPE(op_value), cur_cls, current_method, off)
+                    oth_cls = self.classes[type_info]
+
+                    # FIXME: xref_to does not work here! current_method is wrong, as it is not the target!
+                    # In this case that means, that current_method calls the class oth_class.
+                    # Hence, on xref_to the method info is the calling method not the called one,
+                    # as there is no called method!
+                    # With the _new_instance and _const_class can this be deprecated?
+                    # Removing these does not impact tests
+                    cur_cls.add_xref_to(REF_TYPE(op_value), oth_cls, cur_meth, off)
+                    oth_cls.add_xref_from(REF_TYPE(op_value), cur_cls, cur_meth, off)
+
+                    if op_value == 0x1c:
+                        cur_meth.add_xref_const_class(oth_cls, off)
+                        oth_cls.add_xref_const_class(cur_meth, off)
+                    if op_value == 0x22:
+                        cur_meth.add_xref_new_instance(oth_cls, off)
+                        oth_cls.add_xref_new_instance(cur_meth, off)
 
                 # 2) check for method calls: invoke-* (0x6e ... 0x72), invoke-xxx/range (0x74 ... 0x78)
                 elif (0x6e <= op_value <= 0x72) or (0x74 <= op_value <= 0x78):
                     idx_meth = instruction.get_ref_kind()
                     method_info = instruction.cm.vm.get_cm_method(idx_meth)
-                    if method_info:
-                        class_info = method_info[0]
+                    if not method_info:
+                        log.warning("Could not get method_info "
+                                    "for instruction at {} in method at @{}. "
+                                    "Requested IDX {}".format(off, current_method.get_code_off(), idx_meth))
+                        continue
 
-                        method_item = None
-                        # TODO: should create get_method_descriptor inside Analysis
-                        for vm in self.vms:
-                            method_item = vm.get_method_descriptor(method_info[0], method_info[1], ''.join(method_info[2]))
-                            if method_item:
-                                break
+                    class_info = method_info[0].lstrip(b'[')
+                    if class_info[0] != b'L':
+                        # Need to make sure, that we get class types and not other types
+                        # If another type, like int is used, we simply skip it.
+                        continue
 
-                        if not method_item:
-                            # Seems to be an external class, create it first
-                            # Beware: if not all DEX files are loaded at the time create_xref runs
-                            # you will run into problems!
-                            if method_info[0] not in self.classes:
-                                self.classes[method_info[0]] = ClassAnalysis(ExternalClass(method_info[0]))
-                            method_item = self.classes[method_info[0]].get_fake_method(method_info[1], method_info[2])
+                    # Resolve the second MethodAnalysis
+                    oth_meth = self._resolve_method(class_info, method_info[1], method_info[2])
 
-                        self.classes[cur_cls_name].AddMXrefTo(current_method, self.classes[class_info], method_item, off)
-                        self.classes[class_info].AddMXrefFrom(method_item, self.classes[cur_cls_name], current_method, off)
+                    oth_cls = self.classes[class_info]
 
-                        # Internal xref related to class manipulation
-                        if class_info in self.classes and class_info != cur_cls_name:
-                            self.classes[cur_cls_name].AddXrefTo(REF_TYPE(op_value), self.classes[class_info], method_item, off)
-                            self.classes[class_info].AddXrefFrom(REF_TYPE(op_value), self.classes[cur_cls_name], current_method, off)
+                    # FIXME: we could merge add_method_xref_* and add_xref_*
+                    cur_cls.add_method_xref_to(cur_meth, oth_cls, oth_meth, off)
+                    oth_cls.add_method_xref_from(oth_meth, cur_cls, cur_meth, off)
+                    # Internal xref related to class manipulation
+                    cur_cls.add_xref_to(REF_TYPE(op_value), oth_cls, oth_meth, off)
+                    oth_cls.add_xref_from(REF_TYPE(op_value), cur_cls, cur_meth, off)
 
                 # 3) check for string usage: const-string (0x1a), const-string/jumbo (0x1b)
                 elif 0x1a <= op_value <= 0x1b:
@@ -1240,8 +1542,7 @@ class Analysis:
                     if string_value not in self.strings:
                         self.strings[string_value] = StringAnalysis(string_value)
 
-                    # TODO: The bytecode offset is stored for classes but not here?
-                    self.strings[string_value].AddXrefFrom(self.classes[cur_cls_name], current_method)
+                    self.strings[string_value].add_xref_from(cur_cls, cur_meth, off)
 
                 # TODO maybe we should add a step 3a) here and check for all const fields. You can then xref for integers etc!
                 # But: This does not work, as const fields are usually optimized internally to const calls...
@@ -1251,16 +1552,17 @@ class Analysis:
                     idx_field = instruction.get_ref_kind()
                     field_info = instruction.cm.vm.get_cm_field(idx_field)
                     field_item = instruction.cm.vm.get_field_descriptor(field_info[0], field_info[2], field_info[1])
-                    # TODO: The bytecode offset is stored for classes but not here?
-                    if field_item:
-                        if (0x52 <= op_value <= 0x58) or (0x60 <= op_value <= 0x66):
-                            # read access to a field
-                            self.classes[cur_cls_name].AddFXrefRead(current_method, self.classes[cur_cls_name], field_item)
-                        else:
-                            # write access to a field
-                            self.classes[cur_cls_name].AddFXrefWrite(current_method, self.classes[cur_cls_name], field_item)
+                    if not field_item:
+                        continue
 
-                off += instruction.get_length()
+                    if (0x52 <= op_value <= 0x58) or (0x60 <= op_value <= 0x66):
+                        # read access to a field
+                        self.classes[cur_cls_name].add_field_xref_read(cur_meth, cur_cls, field_item, off)
+                        cur_meth.add_xref_read(cur_cls, field_item, off)
+                    else:
+                        # write access to a field
+                        self.classes[cur_cls_name].add_field_xref_write(cur_meth, cur_cls, field_item, off)
+                        cur_meth.add_xref_write(cur_cls, field_item, off)
 
     def get_method(self, method):
         """
@@ -1269,11 +1571,43 @@ class Analysis:
 
         :param method: :class:`EncodedMethod` to search for
         :return: :class:`MethodAnalysis` object for the given method, or None if method was not found
+        :rtype: MethodAnalysis
         """
         if method in self.methods:
             return self.methods[method]
-        else:
-            return None
+        return None
+
+    # Alias
+    get_method_analysis = get_method
+
+    def _resolve_method(self, class_name, method_name, method_descriptor):
+        """
+        Resolves the Method and returns MethodAnalysis.
+        Will automatically create ExternalMethods if can not resolve and add to the ClassAnalysis etc
+
+        :param str class_name:
+        :param str method_name:
+        :param List[str] method_descriptor: Tuple which has parameters and return type, i.e. ['(I Z)', 'V']
+        :return:
+        :rtype: MethodAnalysis
+        """
+        m_hash = (class_name, method_name, mutf8.MUTF8String.join(method_descriptor))
+        if m_hash not in self.__method_hashes:
+            # Need to create a new method
+            if class_name not in self.classes:
+                # External class? no problem!
+                self.classes[class_name] = ClassAnalysis(ExternalClass(class_name))
+
+            # Create external method
+            meth = ExternalMethod(class_name, method_name, method_descriptor)
+            meth_analysis = MethodAnalysis(None, meth)
+
+            # add to all the collections we have
+            self.__method_hashes[m_hash] = meth_analysis
+            self.classes[class_name].add_method(meth_analysis)
+            self.methods[meth] = meth_analysis
+
+        return self.__method_hashes[m_hash]
 
     def get_method_by_name(self, class_name, method_name, method_descriptor):
         """
@@ -1283,28 +1617,11 @@ class Analysis:
         :param method_name: name of the method, for example 'onCreate'
         :param method_descriptor: descriptor, for example '(I I Ljava/lang/String)V
         :return: :class:`EncodedMethod` or None if method was not found
+        :rtype: androguard.core.bytecodes.dvm.EncodedMethod
         """
-        if class_name in self.classes:
-            for method in self.classes[class_name].get_vm_class().get_methods():
-                if method.get_name() == method_name and method.get_descriptor() == method_descriptor:
-                    return method
-        return None
-
-    def get_method_analysis(self, method):
-        """
-        Returns the crossreferencing object for a given Method.
-
-        Beware: the similar named function :meth:`~get_method()` will return
-        a :class:`MethodAnalysis` object, while this function returns a :class:`MethodClassAnalysis` object!
-
-        This Method will only work after a run of :meth:`~create_xref()`
-
-        :param method: :class:`EncodedMethod`
-        :return: :class:`MethodClassAnalysis` for the given method or None, if method was not found
-        """
-        class_analysis = self.get_class_analysis(method.get_class_name())
-        if class_analysis:
-            return class_analysis.get_method_analysis(method)
+        m_a = self.get_method_analysis_by_name(class_name, method_name, method_descriptor)
+        if m_a and not m_a.is_external():
+            return m_a.get_method()
         return None
 
     def get_method_analysis_by_name(self, class_name, method_name, method_descriptor):
@@ -1317,19 +1634,21 @@ class Analysis:
         :param class_name: name of the class, for example `'Ljava/lang/Object;'`
         :param method_name: name of the method, for example `'onCreate'`
         :param method_descriptor: method descriptor, for example `'(I I)V'`
-        :return: :class:`MethodClassAnalysis`
+        :return: :class:`MethodAnalysis`
+        :rtype: MethodAnalysis
         """
-        method = self.get_method_by_name(class_name, method_name, method_descriptor)
-        if method:
-            return self.get_method_analysis(method)
-        return None
+        m_hash = (class_name, method_name, method_descriptor)
+        if m_hash not in self.__method_hashes:
+            return None
+        return self.__method_hashes[m_hash]
 
     def get_field_analysis(self, field):
         """
         Get the FieldAnalysis for a given fieldname
 
-        :param field: TODO
-        :return: :class:`FieldClassAnalysis`
+        :param androguard.core.bytecodes.dvm.EncodedField field: the field
+        :return: :class:`FieldAnalysis`
+        :rtype: FieldAnalysis
         """
         class_analysis = self.get_class_analysis(field.get_class_name())
         if class_analysis:
@@ -1342,6 +1661,7 @@ class Analysis:
 
         :param class_name: classname like 'Ljava/lang/Object;' (including L and ;)
         :return: True if class was found, False otherwise
+        :rtype: bool
         """
         return class_name in self.classes
 
@@ -1351,6 +1671,7 @@ class Analysis:
 
         :param class_name: classname like 'Ljava/lang/Object;' (including L and ;)
         :return: :class:`ClassAnalysis`
+        :rtype: ClassAnalysis
         """
         return self.classes.get(class_name)
 
@@ -1359,7 +1680,7 @@ class Analysis:
         Returns all external classes, that means all classes that are not
         defined in the given set of `DalvikVMObjects`.
 
-        :rtype: generator of `ClassAnalysis`
+        :rtype: Iterator[ClassAnalysis]
         """
         for cls in self.classes.values():
             if cls.is_external():
@@ -1370,7 +1691,7 @@ class Analysis:
         Returns all external classes, that means all classes that are
         defined in the given set of :class:`~DalvikVMFormat`.
 
-        :rtype: generator of :class:`~ClassAnalysis`
+        :rtype: Iterator[ClassAnalysis]
         """
         for cls in self.classes.values():
             if not cls.is_external():
@@ -1380,7 +1701,7 @@ class Analysis:
         """
         Returns a dictionary of strings and their corresponding :class:`StringAnalysis`
 
-        :return: a dictionary
+        :rtype: Dict[str, StringAnalysis]
         """
         return self.strings
 
@@ -1388,7 +1709,7 @@ class Analysis:
         """
         Returns a list of :class:`StringAnalysis` objects
 
-        :rtype: list of :class:`StringAnalysis`
+        :rtype: Iterator[StringAnalysis]
         """
         return self.strings.values()
 
@@ -1398,23 +1719,24 @@ class Analysis:
 
         Returns both internal and external classes (if any)
 
-        :rtype: list of :class:`ClassAnalysis`
+        :rtype: Iterator[ClassAnalysis]
         """
         return self.classes.values()
 
     def get_methods(self):
         """
-        Returns a list of `MethodClassAnalysis` objects
+        Returns a list of `MethodAnalysis` objects
+
+        :rtype: Iterator[MethodAnalysis]
 
         """
-        for c in self.classes.values():
-            for m in c.get_methods():
-                yield m
+        yield from self.methods.values()
 
     def get_fields(self):
         """
-        Returns a list of `FieldClassAnalysis` objects
+        Returns a list of `FieldAnalysis` objects
 
+        :rtype: Iterator[FieldAnalysis]
         """
         for c in self.classes.values():
             for f in c.get_fields():
@@ -1428,8 +1750,9 @@ class Analysis:
 
         :param name: regular expression for class name (default ".*")
         :param no_external: Remove external classes from the output (default False)
-        :rtype: generator of `ClassAnalysis`
+        :rtype: Iterator[ClassAnalysis]
         """
+        name = bytes(mutf8.MUTF8String.from_str(name))
         for cname, c in self.classes.items():
             if no_external and isinstance(c.get_vm_class(), ExternalClass):
                 continue
@@ -1440,7 +1763,7 @@ class Analysis:
             accessflags=".*", no_external=False):
         """
         Find a method by name using regular expression.
-        This method will return all MethodClassAnalysis objects, which match the
+        This method will return all MethodAnalysis objects, which match the
         classname, methodname, descriptor and accessflags of the method.
 
         :param classname: regular expression for the classname
@@ -1448,8 +1771,11 @@ class Analysis:
         :param descriptor: regular expression for the descriptor
         :param accessflags: regular expression for the accessflags
         :param no_external: Remove external method from the output (default False)
-        :rtype: generator of `MethodClassAnalysis`
+        :rtype: Iterator[MethodAnalysis]
         """
+        classname = bytes(mutf8.MUTF8String.from_str(classname))
+        methodname = bytes(mutf8.MUTF8String.from_str(methodname))
+        descriptor = bytes(mutf8.MUTF8String.from_str(descriptor))
         for cname, c in self.classes.items():
             if re.match(classname, cname):
                 for m in c.get_methods():
@@ -1469,8 +1795,9 @@ class Analysis:
         Find strings by regex
 
         :param string: regular expression for the string to search for
-        :rtype: generator of `StringAnalysis`
+        :rtype: Iterator[StringAnalysis]
         """
+        string = bytes(mutf8.MUTF8String.from_str(string))
         for s, sa in self.strings.items():
             if re.match(string, s):
                 yield sa
@@ -1483,8 +1810,11 @@ class Analysis:
         :param fieldname: regular expression of the fieldname
         :param fieldtype: regular expression of the fieldtype
         :param accessflags: regular expression of the access flags
-        :rtype: generator of `FieldClassAnalysis`
+        :rtype: Iterator[FieldAnalysis]
         """
+        classname = bytes(mutf8.MUTF8String.from_str(classname))
+        fieldname = bytes(mutf8.MUTF8String.from_str(fieldname))
+        fieldtype = bytes(mutf8.MUTF8String.from_str(fieldtype))
         for cname, c in self.classes.items():
             if re.match(classname, cname):
                 for f in c.get_fields():
@@ -1495,7 +1825,7 @@ class Analysis:
                         yield f
 
     def __repr__(self):
-        return "<analysis.Analysis VMs: {}, Classes: {}, Strings: {}>".format(len(self.vms), len(self.classes), len(self.strings))
+        return "<analysis.Analysis VMs: {}, Classes: {}, Methods: {}, Strings: {}>".format(len(self.vms), len(self.classes), len(self.methods), len(self.strings))
 
     def get_call_graph(self, classname=".*", methodname=".*", descriptor=".*",
                        accessflags=".*", no_isolated=False, entry_points=[]):
@@ -1504,56 +1834,83 @@ class Analysis:
         The filters are the same as in
         :meth:`~androguard.core.analaysis.analaysis.Analysis.find_methods`
 
-        A networkx.DiGraph is returned, containing all edges only once!
-        that means, if a method calls some method twice or more often, there will
-        only be a single connection.
+        A networkx.MultiDiGraph is returned, containing all xrefs.
+        That means a method which calls another method multiple times, will have multiple
+        edges between them. Attached to the edge is the attribute `offset`, which gives
+        the code offset inside the method of the call.
+
+        Specifying filters will not remove the methods if they are called by some other method.
+
+        The callgraph will check for both directions of edges. Thus, if you specify a single class
+        as input, it will contain all classes which are called by this class (xref_to),
+        as well as all methods who calls the specified one (xref_from).
+
+        Each node will contain the following meta information as attribute:
+
+        * external: is the method external or not (boolean)
+        * entrypoint: is the method a known entry point (boolean)
+        * native: is the method a native method by signature (boolean)
+        * public: is the method declared public (boolean)
+        * static: is the method declared static (boolean)
+        * vm: An ID of the DEX file where this method is declared or 0 if external (signed int)
+        * codesize: size of code of the method or zero if external (int)
 
         :param classname: regular expression of the classname (default: ".*")
-        :param fieldname: regular expression of the fieldname (default: ".*")
-        :param fieldtype: regular expression of the fieldtype (default: ".*")
+        :param methodname: regular expression of the methodname (default: ".*")
+        :param descriptor: regular expression of the descriptor (default: ".*")
         :param accessflags: regular expression of the access flags (default: ".*")
         :param no_isolated: remove isolated nodes from the graph, e.g. methods which do not call anything (default: False)
         :param entry_points: A list of classes that are marked as entry point
 
-        :rtype: DiGraph
+        :rtype: networkx.MultiDiGraph
         """
 
         def _add_node(G, method):
             """
-            Wrapper to add methods to a graph
-            """
-            if method not in G.node:
-                G.add_node(method,
-                           external=isinstance(method, ExternalMethod),
-                           entrypoint=method.get_class_name() in entry_points,
-                           native="native" in method.get_access_flags_string(),
-                           public="public" in method.get_access_flags_string(),
-                           static="static" in method.get_access_flags_string(),
-                           )
+            Wrapper to add methods to a graph without duplication
 
-        CG = nx.DiGraph()
+            :param nx.MultiDiGraph G:
+            :param MethodAnalysis method:
+            """
+            if method in G.nodes:
+                return
+
+            G.add_node(method,
+                       external=method.is_external(),
+                       entrypoint=method.class_name in entry_points,
+                       native="native" in method.access,
+                       public="public" in method.access,
+                       static="static" in method.access,
+                       vm=hash(method.get_method().CM.vm) if not method.is_external() else 0,
+                       codesize=len(list(method.get_method().get_instructions())) if not method.is_external() else 0,
+                       )
+
+        CG = nx.MultiDiGraph()
 
         # Note: If you create the CG from many classes at the same time, the drawing
-        # will be a total mess...
+        # will be a total mess... Hence it is recommended to reduce the number of nodes beforehand.
+        # Obviously, you can always do this later at the costs of computational power.
         for m in self.find_methods(classname=classname, methodname=methodname,
                                    descriptor=descriptor, accessflags=accessflags):
-            orig_method = m.get_method()
-            log.info("Found Method --> {}".format(orig_method))
+            log.info("Adding Method '{}' to callgraph".format(m.full_name))
 
-            if no_isolated and len(m.get_xref_to()) == 0:
-                log.info("Skipped {}, because if has no xrefs".format(orig_method))
+            if no_isolated and len(m.get_xref_to()) == 0 and len(m.get_xref_from()) == 0:
+                log.info("Skipped {}, because if has no xrefs".format(m.full_name))
                 continue
 
-            _add_node(CG, orig_method)
+            _add_node(CG, m)
 
-            for other_class, callee, offset in m.get_xref_to():
+            for _, callee, offset in m.get_xref_to():
                 _add_node(CG, callee)
+                CG.add_edge(m, callee, key=offset, offset=offset)
 
-                # As this is a DiGraph and we are not interested in duplicate edges,
-                # check if the edge is already in the edge set.
-                # If you need all calls, you probably want to check out MultiDiGraph
-                if not CG.has_edge(orig_method, callee):
-                    CG.add_edge(orig_method, callee)
+            for _, caller, offset in m.get_xref_from():
+                # If _all_ methods are added to the CG, this will not make any difference.
+                # But, if only a single class is chosen as a seed point, we require this information too!
+                # This is particularly useful for external classes, as they do not have xref_to,
+                # thus if an external class is chosen as starting point, it will generate an empty graph.
+                _add_node(CG, caller)
+                CG.add_edge(caller, m, key=offset, offset=offset)
 
         return CG
 
@@ -1571,8 +1928,8 @@ class Analysis:
         As Strings can contain nearly anything, use :meth:`find_strings` instead.
 
         * Each `CLASS_` item will return a :class:`~ClassAnalysis`
-        * Each `METHOD_` item will return a :class:`~MethodClassAnalysis`
-        * Each `FIELD_` item will return a :class:`~FieldClassAnalysis`
+        * Each `METHOD_` item will return a :class:`~MethodAnalysis`
+        * Each `FIELD_` item will return a :class:`~FieldAnalysis`
         """
         # TODO: it would be fun to have the classes organized like the packages. I.e. you could do dx.CLASS_xx.yyy.zzz
         for cls in self.get_classes():
@@ -1630,7 +1987,7 @@ class Analysis:
             If the wrong API level is used, the results might be wrong.
 
         :param apilevel: API level to load, or None for default
-        :return: yields tuples of :class:`MethodClassAnalysis` (of the API method) and list of permission string
+        :return: yields tuples of :class:`MethodAnalysis` (of the API method) and list of permission string
         """
 
         # TODO maybe have the API level loading in the __init__ method and pass the APK as well?
@@ -1664,7 +2021,7 @@ class Analysis:
 
         :param permission: the name of the android permission (usually 'android.permission.XXX')
         :param apilevel: the requested API level or None for default
-        :return: yields :class:`MethodClassAnalysis` objects for all using API methods
+        :return: yields :class:`MethodAnalysis` objects for all using API methods
         """
 
         # TODO maybe have the API level loading in the __init__ method and pass the APK as well?
@@ -1684,14 +2041,27 @@ class Analysis:
                 if meth.permission_api_name in apis:
                     yield meth_analysis
 
+    def get_android_api_usage(self):
+        """
+        Get all usage of the Android APIs inside the Analysis.
+
+        :return: yields :class:`MethodAnalysis` objects for all Android APIs methods
+        """
+
+        for cls in self.get_external_classes():
+            for meth_analysis in cls.get_methods():
+                if meth_analysis.is_android_api():
+                    yield meth_analysis
+
 
 def is_ascii_obfuscation(vm):
     """
     Tests if any class inside a DalvikVMObject
     uses ASCII Obfuscation (e.g. UTF-8 Chars in Classnames)
 
-    :param vm: `DalvikVMObject`
+    :param androguard.core.bytecodes.dvm.DalvikVMFormat vm: `DalvikVMObject`
     :return: True if ascii obfuscation otherwise False
+    :rtype: bool
     """
     for classe in vm.get_classes():
         if is_ascii_problem(classe.get_name()):
