@@ -3,7 +3,8 @@ import unittest
 
 import sys
 
-from androguard.core.bytecodes import dvm, mutf8
+from androguard.core import  mutf8
+from androguard.core.bytecodes import dvm
 from androguard.core.analysis import analysis
 
 
@@ -27,25 +28,20 @@ class StringTest(unittest.TestCase):
                 self.assertIn(s, d.get_strings())
 
     def testMUTF8(self):
-        self.assertEqual("\x67", mutf8.decode(b"\x67"))
-        # Null byte
-        self.assertEqual("\x00", mutf8.decode(b"\xc0\x80"))
-        self.assertEqual("\uacf0", mutf8.decode(b"\xea\xb3\xb0"))
-        # Surrogates
-        self.assertEqual("\ud83d\ude4f", mutf8.decode(b"\xed\xa0\xbd\xed\xb9\x8f"))
-        self.assertEqual("\ud853\udf5c", mutf8.decode(b"\xed\xa1\x93\xed\xbd\x9c"))
-        # Lonely surrogates
-        self.assertEqual("\ud853", mutf8.decode(b"\xed\xa1\x93"))
-        self.assertEqual("\udf5c", mutf8.decode(b"\xed\xbd\x9c"))
-        # Normal ASCII String
-        self.assertEqual("hello world", mutf8.decode(b"\x68\x65\x6c\x6c\x6f\x20\x77\x6f\x72\x6c\x64"))
+        # self.assertEqual("\x67", mutf8.decode(b"\x67"))
+        # # Null byte
+        # self.assertEqual("\x00", mutf8.decode(b"\xc0\x80"))
+        # self.assertEqual("\uacf0", mutf8.decode(b"\xea\xb3\xb0"))
+        # # Surrogates
+        # self.assertEqual("\ud83d\ude4f", mutf8.decode(b"\xed\xa0\xbd\xed\xb9\x8f"))
+        # self.assertEqual("\ud853\udf5c", mutf8.decode(b"\xed\xa1\x93\xed\xbd\x9c"))
+        # # Lonely surrogates
+        # self.assertEqual("\ud853", mutf8.decode(b"\xed\xa1\x93"))
+        # self.assertEqual("\udf5c", mutf8.decode(b"\xed\xbd\x9c"))
+        # # Normal ASCII String
+        # self.assertEqual("hello world", mutf8.decode(b"\x68\x65\x6c\x6c\x6f\x20\x77\x6f\x72\x6c\x64"))
 
-        # Test the patching of strings
-
-        self.assertEqual("hello world", mutf8.patch_string(mutf8.decode(b"\x68\x65\x6c\x6c\x6f\x20\x77\x6f\x72\x6c\x64")))
-        self.assertEqual("\U00024f5c", mutf8.patch_string(mutf8.decode(b"\xed\xa1\x93\xed\xbd\x9c")))
-        self.assertEqual("\U0001f64f", mutf8.patch_string(mutf8.decode(b"\xed\xa0\xbd\xed\xb9\x8f")))
-        self.assertEqual("\\ud853", mutf8.patch_string(mutf8.decode(b"\xed\xa1\x93")))
+        # Testing decode
 
         b = b"\xed\xa1\x93\xed\xbd\x9c" + \
             b"\xed\xa0\xbd\xed\xb9\x8f" + \
@@ -54,8 +50,21 @@ class StringTest(unittest.TestCase):
             b"\x68\x65\x6c\x6c\x6f\x20\x77\x6f\x72\x6c\x64" + \
             b"\xc0\x80"
 
-        self.assertEqual("\U00024f5c\U0001f64f\\ud83d\uacf0hello world\x00", mutf8.patch_string(mutf8.decode(b)))
+        self.assertEqual("hello world", mutf8.decode(b"\x68\x65\x6c\x6c\x6f\x20\x77\x6f\x72\x6c\x64").encode('utf8', errors='backslashreplace').decode('utf8'))
+        self.assertEqual("\U00024f5c", mutf8.decode(b"\xed\xa1\x93\xed\xbd\x9c").encode('utf8', errors='backslashreplace').decode('utf8'))
+        self.assertEqual("\U0001f64f", mutf8.decode(b"\xed\xa0\xbd\xed\xb9\x8f").encode('utf8', errors='backslashreplace').decode('utf8'))
+        self.assertEqual("\\ud853", mutf8.decode(b"\xed\xa1\x93").encode('utf8', errors='backslashreplace').decode('utf8'))
+        self.assertEqual("\U00024f5c\U0001f64f\\ud83d\uacf0hello world\x00", mutf8.decode(b).encode('utf8', errors='backslashreplace').decode('utf8'))
 
+        # Testing encode
+
+        self.assertEqual(b"\x68\x65\x6c\x6c\x6f\x20\x77\x6f\x72\x6c\x64", mutf8.encode("hello world"))
+        self.assertEqual(b"\xed\xa1\x93\xed\xbd\x9c", mutf8.encode("\U00024f5c"))
+        self.assertEqual(b"\xed\xa0\xbd\xed\xb9\x8f", mutf8.encode("\U0001f64f"))
+        self.assertEqual(b"\xed\xa1\x93", mutf8.encode("\ud853"))
+        self.assertEqual(b, mutf8.encode("\U00024f5c\U0001f64f\ud83d\uacf0hello world\x00"))
+
+        self.assertEqual(mutf8.MUTF8String(b), mutf8.MUTF8String.from_str("\U00024f5c\U0001f64f\ud83d\uacf0hello world\x00"))
 
 
 if __name__ == '__main__':
